refactor instructor profile edit page into components
